move actions to state-badge
