+++ src/panels/lovelace/entity-rows/hui-media-player-entity-row.ts
@@ -24,8 +24,6 @@ import {
   SUPPORT_PAUSE,
 } from "../../../data/media-player";
 import { hasConfigOrEntityChanged } from "../common/has-changed";
-import { longPress } from "../common/directives/long-press-directive";
-import { handleClick } from "../common/handle-click";
 
 @customElement("hui-media-player-entity-row")
 class HuiMediaPlayerEntityRow extends LitElement implements EntityRow {
@@ -66,11 +64,8 @@ class HuiMediaPlayerEntityRow extends LitElement implements EntityRow {
 
     return html`
       <hui-generic-entity-row
-        @ha-click=${this._handleTap}
-        @ha-hold=${this._handleHold}
-        .longPress=${longPress()}
-        .hass=${this.hass}
-        .config=${this._config}
+        .hass="${this.hass}"
+        .config="${this._config}"
         .showSecondary="false"
       >
         ${OFF_STATES.includes(stateObj.state)
@@ -162,14 +157,6 @@ class HuiMediaPlayerEntityRow extends LitElement implements EntityRow {
     this.hass!.callService("media_player", "media_next_track", {
       entity_id: this._config!.entity,
     });
-  }
-
-  private _handleTap() {
-    handleClick(this, this.hass!, this._config!, false);
-  }
-
-  private _handleHold() {
-    handleClick(this, this.hass!, this._config!, true);
   }
 }
