+++ src/panels/lovelace/entity-rows/hui-script-entity-row.ts
@@ -16,8 +16,6 @@ import "../components/hui-warning";
 import { HomeAssistant } from "../../../types";
 import { EntityRow, EntityConfig } from "./types";
 import { hasConfigOrEntityChanged } from "../common/has-changed";
-import { longPress } from "../common/directives/long-press-directive";
-import { handleClick } from "../common/handle-click";
 
 @customElement("hui-script-entity-row")
 class HuiScriptEntityRow extends LitElement implements EntityRow {
@@ -56,13 +54,7 @@ class HuiScriptEntityRow extends LitElement implements EntityRow {
     }
 
     return html`
-      <hui-generic-entity-row
-        @ha-click=${this._handleTap}
-        @ha-hold=${this._handleHold}
-        .longPress=${longPress()}
-        .hass=${this.hass}
-        .config=${this._config}
-      >
+      <hui-generic-entity-row .hass="${this.hass}" .config="${this._config}">
         ${stateObj.attributes.can_cancel
           ? html`
               <ha-entity-toggle
@@ -92,14 +84,6 @@ class HuiScriptEntityRow extends LitElement implements EntityRow {
     this.hass!.callService("script", "turn_on", {
       entity_id: this._config!.entity,
     });
-  }
-
-  private _handleTap() {
-    handleClick(this, this.hass!, this._config!, false);
-  }
-
-  private _handleHold() {
-    handleClick(this, this.hass!, this._config!, true);
   }
 }
