+++ src/panels/lovelace/entity-rows/hui-sensor-entity-row.ts
@@ -17,8 +17,6 @@ import { HomeAssistant } from "../../../types";
 import { EntityRow, EntityConfig } from "./types";
 import { hasConfigOrEntityChanged } from "../common/has-changed";
 import { computeStateDisplay } from "../../../common/entity/compute_state_display";
-import { longPress } from "../common/directives/long-press-directive";
-import { handleClick } from "../common/handle-click";
 
 interface SensorEntityConfig extends EntityConfig {
   format?: "relative" | "date" | "time" | "datetime";
@@ -61,13 +59,7 @@ class HuiSensorEntityRow extends LitElement implements EntityRow {
     }
 
     return html`
-      <hui-generic-entity-row
-        @ha-click=${this._handleTap}
-        @ha-hold=${this._handleHold}
-        .longPress=${longPress()}
-        .hass=${this.hass}
-        .config=${this._config}
-      >
+      <hui-generic-entity-row .hass="${this.hass}" .config="${this._config}">
         <div>
           ${stateObj.attributes.device_class === "timestamp" &&
           stateObj.state !== "unavailable"
@@ -94,14 +86,6 @@ class HuiSensorEntityRow extends LitElement implements EntityRow {
         text-align: right;
       }
     `;
-  }
-
-  private _handleTap() {
-    handleClick(this, this.hass!, this._config!, false);
-  }
-
-  private _handleHold() {
-    handleClick(this, this.hass!, this._config!, true);
   }
 }
