+++ src/panels/lovelace/entity-rows/hui-timer-entity-row.ts
@@ -17,8 +17,6 @@ import { HomeAssistant } from "../../../types";
 import { EntityConfig } from "./types";
 import { HassEntity } from "home-assistant-js-websocket";
 import { hasConfigOrEntityChanged } from "../common/has-changed";
-import { longPress } from "../common/directives/long-press-directive";
-import { handleClick } from "../common/handle-click";
 
 @customElement("hui-timer-entity-row")
 class HuiTimerEntityRow extends LitElement {
@@ -72,13 +70,7 @@ class HuiTimerEntityRow extends LitElement {
     }
 
     return html`
-      <hui-generic-entity-row
-        @ha-click=${this._handleTap}
-        @ha-hold=${this._handleHold}
-        .longPress=${longPress()}
-        .hass=${this.hass}
-        .config=${this._config}
-      >
+      <hui-generic-entity-row .hass="${this.hass}" .config="${this._config}">
         <div>${this._computeDisplay(stateObj)}</div>
       </hui-generic-entity-row>
     `;
@@ -149,14 +141,6 @@ class HuiTimerEntityRow extends LitElement {
     }
 
     return display;
-  }
-
-  private _handleTap() {
-    handleClick(this, this.hass!, this._config!, false);
-  }
-
-  private _handleHold() {
-    handleClick(this, this.hass!, this._config!, true);
   }
 }
