+++ src/panels/lovelace/entity-rows/hui-toggle-entity-row.ts
@@ -15,8 +15,6 @@ import { computeStateDisplay } from "../../../common/entity/compute_state_displa
 import { HomeAssistant } from "../../../types";
 import { EntityRow, EntityConfig } from "./types";
 import { hasConfigOrEntityChanged } from "../common/has-changed";
-import { longPress } from "../common/directives/long-press-directive";
-import { handleClick } from "../common/handle-click";
 
 @customElement("hui-toggle-entity-row")
 class HuiToggleEntityRow extends LitElement implements EntityRow {
@@ -55,13 +53,7 @@ class HuiToggleEntityRow extends LitElement implements EntityRow {
     }
 
     return html`
-      <hui-generic-entity-row
-        @ha-click=${this._handleTap}
-        @ha-hold=${this._handleHold}
-        .longPress=${longPress()}
-        .hass=${this.hass}
-        .config=${this._config}
-      >
+      <hui-generic-entity-row .hass="${this.hass}" .config="${this._config}">
         ${stateObj.state === "on" || stateObj.state === "off"
           ? html`
               <ha-entity-toggle
@@ -80,14 +72,6 @@ class HuiToggleEntityRow extends LitElement implements EntityRow {
             `}
       </hui-generic-entity-row>
     `;
-  }
-
-  private _handleTap() {
-    handleClick(this, this.hass!, this._config!, false);
-  }
-
-  private _handleHold() {
-    handleClick(this, this.hass!, this._config!, true);
   }
 }
